/Candles return bars

+++ db/dbstore/inf.go
@@ -9,8 +9,8 @@ import (
 )
 
 type ForexCandles interface {
-	UpsertCandles(ctx context.Context, symbol_id int, resolution models.Resolution, bars *models.Candles) *errors.AppError
-	QueryCandles(ctx context.Context, symbol_id int, resolution models.Resolution, from time.Time, exclusiveTo time.Time, limit int64) (*models.Candles, *errors.AppError)
+	UpsertCandles(ctx context.Context, symbol_id int64, resolution models.Resolution, bars *models.Candles) *errors.AppError
+	QueryCandles(ctx context.Context, symbol_id int64, resolution models.Resolution, from time.Time, exclusiveTo time.Time, limit int64) (*models.Candles, *errors.AppError)
 }
 
 type DBStore interface {
